+++ next-tavla/src/Admin/scenarios/Edit/components/AddTile/index.tsx
@@ -2,9 +2,9 @@ import { Button } from '@entur/button'
 import { Dropdown, MultiSelect, SearchableDropdown } from '@entur/dropdown'
 import { SearchIcon } from '@entur/icons'
 import { Heading1 } from '@entur/typography'
-import { useStopPlaceSearch } from './hooks/useStopPlaceSearch'
-import { useCountiesSearch } from './hooks/useCountiesSearch'
-import { useQuaySearch } from './hooks/useQuaySearch'
+import { useStopPlaceSearch } from '../../hooks/useStopPlaceSearch'
+import { useCountiesSearch } from '../../hooks/useCountiesSearch'
+import { useQuaySearch } from '../../hooks/useQuaySearch'
 import { useEditSettingsDispatch } from '../../utils/contexts'
 import { useToast } from '@entur/alert'
